Shortcut functions moved to a single file
