+++ frontend-html/src/model/actions-ui/DataView/TableView/onFieldKeyDown.ts
@@ -28,13 +28,13 @@ import { handleError } from "model/actions/handleError";
 import { getDataView } from "model/selectors/DataView/getDataView";
 import { shouldProceedToChangeRow } from "model/actions-ui/DataView/TableView/shouldProceedToChangeRow";
 import { getGridFocusManager } from "model/entities/GridFocusManager";
-import { isSaveShortcut } from "utils/keyShortcuts";
 import {
+  isSaveShortcut,
   isAddRecordShortcut,
   isDeleteRecordShortcut,
   isDuplicateRecordShortcut,
   isFilterRecordShortcut
-} from "utils/shortcuts";
+} from "utils/keyShortcuts";
 import { onDeleteRowClick } from "model/actions-ui/DataView/onDeleteRowClick";
 import { onCreateRowClick } from "model/actions-ui/DataView/onCreateRowClick";
 import { onCopyRowClick } from "model/actions-ui/DataView/onCopyRowClick";
@@ -52,11 +52,6 @@ export function onFieldKeyDown(ctx: any) {
       const dataView = getDataView(ctx);
       const tablePanelView = getTablePanelView(ctx);
       tablePanelView.handleEditorKeyDown(event);
-      if( isSaveShortcut(event)){
-        tablePanelView.setEditing(false);
-        yield*flushCurrentRowData(ctx)();
-        return;
-      }
       switch (event.key) {
         case "Tab": {
           if (isGoingToChangeRow(event)) {
@@ -128,19 +123,20 @@ export function onFieldKeyDown(ctx: any) {
           tablePanelView.triggerOnFocusTable();
           break;
         }
-        default:
-          if(isAddRecordShortcut(event)){
+        default: {
+          if (isSaveShortcut(event)) {
+            tablePanelView.setEditing(false);
+            yield*flushCurrentRowData(ctx)();
+          } else if (isAddRecordShortcut(event)) {
             yield onCreateRowClick(dataView)(event);
-          }
-          else if(isDeleteRecordShortcut(event) ){
+          } else if (isDeleteRecordShortcut(event)) {
             yield onDeleteRowClick(dataView)(event);
-          }
-          else if(isDuplicateRecordShortcut(event)){
+          } else if (isDuplicateRecordShortcut(event)) {
             yield onCopyRowClick(dataView)(event);
-          }
-          else if(isFilterRecordShortcut(event)){
+          } else if (isFilterRecordShortcut(event)) {
             yield onFilterButtonClick(dataView)(event);
           }
+        }
       }
     } catch (e) {
       yield*handleError(ctx)(e);
